exomiser-core/src/main/java/de/charite/compbio/exomiser/core/writer/OutputFormat.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package de.charite.compbio.exomiser.core.writer;

/**
 * Enum for representing the desired format of the output.
 *
 * @author Jules Jacobsen <jules.jacobsen@sanger.ac.uk>
 */
public enum OutputFormat {

    HTML("html"), VCF("vcf"), TSV("tsv");

    private final String fileExtension;

    private OutputFormat(String fileExtension) {
        this.fileExtension = fileExtension;
    }

    public String getFileExtension() {
        return fileExtension;
    }
    
}
